turn and messages have been moved into the avatar from the world and all seems...

See merge request GrumpyGameDev/fsharp/seafarersofsplorr!8

+++ Splorr.Seafarers.Services/Island.fs
@@ -22,7 +22,7 @@ module Island =
         | None ->
             "????"
     
-    let AddVisit (turn: uint32) (island:Island) : Island =
+    let AddVisit (turn: float) (island:Island) : Island =
         match island.VisitCount, island.LastVisit with
         | None, None ->
             {island with
